--- java/com/lets/domain/likePost/LikePostStatus.java
@@ -1,5 +1,0 @@
-package com.lets.domain.likePost;
-
-public enum LikePostStatus {
-    ACTIVE, INACTIVE
-}
