--- java/com/lets/domain/post/PostStatus.java
@@ -1,5 +1,0 @@
-package com.lets.domain.post;
-
-public enum PostStatus {
-    RECRUITING, DONE
-}
