+++ src/main/java/com/lets/springproj/domain/LikePost.java
@@ -1,7 +1,12 @@
 package com.lets.springproj.domain;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
 
+@Getter
+@NoArgsConstructor
 @Entity
 public class LikePost {
     @Id
@@ -23,6 +28,8 @@ public class LikePost {
     private LikePost(Member member, Post post){
         this.member = member;
         this.post = post;
+        status = LikePostStatus.INACTIVE;
+        post.addView();
     }
 
     public static LikePost createLikePost(Member member, Post post){
@@ -34,12 +41,18 @@ public class LikePost {
     }
     //==좋아요 메서드==//
     public void activateLike(){
+        if(this.status == LikePostStatus.ACTIVE){
+            throw new IllegalStateException("이미 좋아요 한 글은 좋아요가 불가능합니다.");
+        }
         this.status = LikePostStatus.ACTIVE;
         post.addLike();
     }
 
     //==좋아요 취소 메서드==//
     public void deactivateLike(){
+        if(this.status == LikePostStatus.INACTIVE) {
+            throw new IllegalStateException("이미 좋아요 취소 한 글은 좋아요 취소가 불가능합니다.");
+        }
         this.status = LikePostStatus.INACTIVE;
         post.minusLike();
     }
