+++ src/main/java/com/lets/springproj/domain/Member.java
@@ -1,9 +1,15 @@
 package com.lets.springproj.domain;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
 import javax.persistence.*;
+import javax.validation.constraints.NotEmpty;
 import java.util.ArrayList;
 import java.util.List;
 
+@Getter
+@NoArgsConstructor
 @Entity
 public class Member {
     @Id @GeneratedValue
@@ -11,6 +17,7 @@ public class Member {
     private Long id;
 
     @Column(nullable = false)
+    @NotEmpty
     private String name;
 
     @OneToMany(mappedBy = "member")
@@ -40,7 +47,7 @@ public class Member {
     //==연관관계 메서드==//
     public void addMemberTechStack(MemberTechStack memberTechStack){
 
-        this.addMemberTechStack(memberTechStack);
+        this.memberTechStacks.add(memberTechStack);
         memberTechStack.setMember(this);
     }
