+++ src/main/java/com/lets/springproj/domain/MemberTechStack.java
@@ -1,10 +1,12 @@
 package com.lets.springproj.domain;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
 @Getter
+@NoArgsConstructor
 @Entity
 public class MemberTechStack {
     @Id @GeneratedValue
