+++ src/main/java/com/lets/springproj/domain/Tag.java
@@ -1,12 +1,15 @@
 package com.lets.springproj.domain;
 
-import lombok.Setter;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import javax.validation.constraints.NotEmpty;
 import java.util.ArrayList;
 import java.util.List;
 
-@Setter
+@Getter
+@NoArgsConstructor
 @Entity
 public class Tag {
     @Id @GeneratedValue
@@ -14,10 +17,11 @@ public class Tag {
     private Long id;
 
     @Column(nullable = false)
+    @NotEmpty
     private String name;
 
     @OneToMany(mappedBy = "tag")
-    private List<PostTechStack> posts = new ArrayList<>();
+    private List<PostTechStack> postTechStacks = new ArrayList<>();
 
     private Tag(String name){
         this.name = name;
@@ -30,6 +34,6 @@ public class Tag {
 
     //==연관관계 메서드==//
     public void addPostTechStack(PostTechStack postTechStack){
-        this.posts.add(postTechStack);
+        this.postTechStacks.add(postTechStack);
     }
 }
